@@ -6,15 +6,17 @@
     "@fortawesome/fontawesome-svg-core": "^1.3.0",
     "@fortawesome/free-regular-svg-icons": "^6.0.0",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "@react-google-maps/api": "^2.7.0",
+    "@fortawesome/react-fontawesome": "^0.1.18",
+    "@react-google-maps/api": "^2.10.2",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "env-cmd": "^10.1.0",
-    "react": "^17.0.2",
+    "leaflet": "^1.8.0",
+    "react": "^18.1.0",
     "react-canvas-confetti": "^1.3.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
+    "react-leaflet": "^4.0.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "react-transition-group": "^4.4.2",
